Add NumberInput component
